connecting the backend to the html
