+++ views/layouts/main.handlebars
@@ -17,7 +17,7 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
         integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
 
-    <link rel="stylesheet" href="./public/stylesheets/style.css" type="text/css" />
+    <link rel="stylesheet" href="public/stylesheets/style.css" type="text/css" />
 
     <script src="js/login.js"></script>
