+++ views/members.handlebars
@@ -4,11 +4,11 @@
     </div>
 </div>
 <div class="container-button">
-    <button type="button" class="btn btn-primary btn-lg">Large button</button>
+    <button type="button" class="btn btn-primary btn-lg"><a class="butt" href="/searchData">Search Data</a></button>
 
-    <button type="button" class="btn btn-primary btn-lg">Large button</button>
+    <button type="button" class="btn btn-primary btn-lg"><a class="butt"  href="/addData">Enter Data</a></button>
 
-    <button type="button" class="btn btn-primary btn-lg">Large button</button>
+    <button type="button" class="btn btn-primary btn-lg"><a class="butt"  href="/newSystem">Add a System</a></button>
 
 
 </div>
