+++ views/searchData.handlebars
@@ -58,7 +58,7 @@
 
 
 
-    <script src="../js/" async defer></script>
+    <script src="public/js/searchData.js" async defer></script>
 </body>
 
 </html>
